Move sql to trust model
