+++ server/services/TrustService.js
@@ -1,14 +1,15 @@
 const expect = require('expect-runtime');
-const knex = require('../../server/database/knex');
 const HttpError = require("../utils/HttpError");
 const EntityModel = require("../models/EntityModel");
+const TrustModel = require("../models/TrustModel");
 
-class TrustModel{
-  async get(){
-    //const trust_relationship_instance = new trust_relationship(1);
-    const list = await knex.select()
-      .table("wallets.entity_trust");
-    return list;
+class TrustService{
+  constructor(){
+    this.trustModel = new TrustModel();
+  }
+
+  getTrustModel(){
+    return this.trustModel;
   }
 
   async request(requestType, walletName){
@@ -20,33 +21,11 @@ class TrustModel{
     //get wallet id
     const entityModel = new EntityModel();
     const wallet = await entityModel.getEntityByWalletName((walletName));
-    await knex("wallets.entity_trust").insert({
+    await this.trustModel.create({
       request_type: requestType,
       target_entity_id: wallet.id,
     });
   }
 }
 
-TrustModel.ENTITY_TRUST_TYPE = {
-  send: 'send',
-  manage: 'manage',
-  deduct: 'deduct',
-}
-
-TrustModel.ENTITY_TRUST_STATE_TYPE = {
-  requested: 'requested',
-  cancelled_by_originator: 'cancelled_by_originator',
-  canceled_by_actor: 'cancelled_by_actor',
-  trusted: 'trusted',
-}
-
-TrustModel.ENTITY_TRUST_REQUEST_TYPE = {
-  send: 'send',
-  received: 'received',
-  manage: 'manage',
-  yield: 'yield',
-  deduct: 'deduct',
-  release: 'release',
-}
-
-module.exports = TrustModel;
+module.exports = TrustService;
